requirements.txt
pytz

pyserial

#influxdb
influxdb_client 